+++ Sources/WaitMerge/OhluluExtensions.swift
@@ -8,32 +8,6 @@
 
 import UIKit
 
-public extension UIStackView {
-    
-    func addArrangedSubviews(_ views: [UIView]) {
-        for view in views {
-            addArrangedSubview(view)
-        }
-    }
-    
-    func forceRemoveAllArrangedSubviews() {
-        for view in arrangedSubviews {
-            removeArrangedSubview(view)
-            view.removeFromSuperview()
-        }
-    }
-    
-    func forceRemoveArrangedSubviews(_ views: [UIView]) {
-        for view in views {
-            if !subviews.contains(view) {
-                continue
-            }
-            removeArrangedSubview(view)
-            view.removeFromSuperview()
-        }
-    }
-}
-
 public extension UIViewController {
     
     @objc func dismiss(_ animated: Bool = true, _ completion: (() -> Void)? = nil) {
